fix horizontal strips

+++ header/gol.h
@@ -9,4 +9,4 @@
 
 
 void rand_gol(unsigned char*, unsigned int);
-void play_gol(unsigned char*, size_t, int);
+void play_gol(unsigned char*, size_t, size_t);
